cripto_seguridad/requirements.txt
Flask
Flask-Mail
mysql-connector-python
opencv-python
numpy
web3
pillow
python-dotenv
Flask-Cors
qrcode
pyzbar
